moving syscall.go to pkg/syscall
